pyproject.toml
[project]
name = "kandidatexamensarbete"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.86.0",
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
